Move executor to IJeiHelpers
