+++ CommonApi/src/main/java/mezz/jei/api/helpers/IJeiHelpers.java
@@ -7,6 +7,7 @@ import mezz.jei.api.runtime.IIngredientManager;
 import net.minecraft.resources.ResourceLocation;
 
 import java.util.Optional;
+import java.util.concurrent.Executor;
 
 /**
  * {@link IJeiHelpers} provides helpers and tools for addon mods.
@@ -67,4 +68,9 @@ public interface IJeiHelpers {
 	 * @since 11.5.0
 	 */
 	IIngredientManager getIngredientManager();
+
+	/**
+	 * Get access to the client executor, which budgets running background tasks on the main thread.
+	 */
+	Executor getClientExecutor();
 }
