+++ Gui/src/main/java/mezz/jei/gui/startup/JeiGuiStarter.java
@@ -115,7 +115,7 @@ public class JeiGuiStarter {
             modIdHelper,
             ingredientVisibility,
             colorHelper,
-            registration.getClientExecutor()
+            jeiHelpers.getClientExecutor()
         );
         ingredientManager.registerIngredientListener(ingredientFilter);
         ingredientVisibility.registerListener(ingredientFilter::onIngredientVisibilityChanged);
