+++ Library/src/main/java/mezz/jei/library/load/registration/RuntimeRegistration.java
@@ -17,8 +17,6 @@ import mezz.jei.library.gui.IngredientListOverlayDummy;
 import mezz.jei.library.gui.recipes.RecipesGuiDummy;
 import mezz.jei.library.ingredients.IngredientFilterApiDummy;
 
-import java.util.concurrent.Executor;
-
 public class RuntimeRegistration implements IRuntimeRegistration {
     private final IRecipeManager recipeManager;
     private final IJeiHelpers jeiHelpers;
@@ -27,7 +25,6 @@ public class RuntimeRegistration implements IRuntimeRegistration {
     private final IIngredientVisibility ingredientVisibility;
     private final IRecipeTransferManager recipeTransferManager;
     private final IScreenHelper screenHelper;
-    private final Executor clientExecutor;
 
     private IIngredientListOverlay ingredientListOverlay = IngredientListOverlayDummy.INSTANCE;
     private IBookmarkOverlay bookmarkOverlay = BookmarkOverlayDummy.INSTANCE;
@@ -41,8 +38,7 @@ public class RuntimeRegistration implements IRuntimeRegistration {
         IIngredientManager ingredientManager,
         IIngredientVisibility ingredientVisibility,
         IRecipeTransferManager recipeTransferManager,
-        IScreenHelper screenHelper,
-        Executor clientExecutor
+        IScreenHelper screenHelper
     ) {
         this.recipeManager = recipeManager;
         this.jeiHelpers = jeiHelpers;
@@ -51,7 +47,6 @@ public class RuntimeRegistration implements IRuntimeRegistration {
         this.ingredientVisibility = ingredientVisibility;
         this.recipeTransferManager = recipeTransferManager;
         this.screenHelper = screenHelper;
-        this.clientExecutor = clientExecutor;
     }
 
     @Override
@@ -123,10 +118,5 @@ public class RuntimeRegistration implements IRuntimeRegistration {
 
     public IIngredientFilter getIngredientFilter() {
         return this.ingredientFilter;
-    }
-
-    @Override
-    public Executor getClientExecutor() {
-        return this.clientExecutor;
     }
 }
