+++ Library/src/main/java/mezz/jei/library/runtime/JeiHelpers.java
@@ -18,6 +18,7 @@ import javax.annotation.Nullable;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Optional;
+import java.util.concurrent.Executor;
 
 public class JeiHelpers implements IJeiHelpers {
 	private final GuiHelper guiHelper;
@@ -27,6 +28,7 @@ public class JeiHelpers implements IJeiHelpers {
 	private final IColorHelper colorHelper;
 	private final IIngredientManager ingredientManager;
 	private final IPlatformFluidHelper<?> platformFluidHelper;
+	private final Executor clientExecutor;
 	private @Nullable Collection<IRecipeCategory<?>> recipeCategories;
 
 	public JeiHelpers(
@@ -35,7 +37,8 @@ public class JeiHelpers implements IJeiHelpers {
 		IModIdHelper modIdHelper,
 		IFocusFactory focusFactory,
 		IColorHelper colorHelper,
-		IIngredientManager ingredientManager
+		IIngredientManager ingredientManager,
+		Executor clientExecutor
 	) {
 		this.guiHelper = guiHelper;
 		this.stackHelper = stackHelper;
@@ -44,6 +47,7 @@ public class JeiHelpers implements IJeiHelpers {
 		this.colorHelper = colorHelper;
 		this.ingredientManager = ingredientManager;
 		this.platformFluidHelper = Services.PLATFORM.getFluidHelper();
+		this.clientExecutor = clientExecutor;
 	}
 
 	public void setRecipeCategories(Collection<IRecipeCategory<?>> recipeCategories) {
@@ -93,5 +97,10 @@ public class JeiHelpers implements IJeiHelpers {
 	@Override
 	public IIngredientManager getIngredientManager() {
 		return ingredientManager;
+	}
+
+	@Override
+	public Executor getClientExecutor() {
+		return clientExecutor;
 	}
 }
